fix unsigned short type for index -> fix zjmp in zork

+++ vm/include/vm.h
@@ -237,7 +237,8 @@ t_ex_ret		read_error(char *filename);
 ** Conversions
 */
 uint32_t		str_to_uint32(char *str);
-uint32_t		memvalue_to_uint32(t_memcase *vm_mem, int index, int size);
+uint32_t		memvalue_to_uint32(t_memcase *vm_mem, unsigned short index,
+									int size);
 
 /*
 ** Parsing
@@ -306,7 +307,7 @@ void			getkey(t_vm *vm);
 t_ex_ret		usage_ret_err(void);
 void			exit_malloc_err(void);
 t_bool			flag_is_applied(int flag, t_vm *vm);
-int				get_mem_index(int index);
+int				get_mem_index(unsigned short index);
 t_bool			is_valid_reg_index(int index);
 
 #endif
